requirements.txt
# Django & DB
Django>=5.2,<6.0
psycopg2-binary>=2.9

# REST API & 서버
djangorestframework
gunicorn
whitenoise

# AI / RAG
transformers>=4.42.0
torch>=2.3.0
langchain>=0.2.0
langchain-community>=0.2.0
openai>=1.35.0
tiktoken
ragas
langchain-postgres

# Vector DB
pgvector   # PostgreSQL에서 벡터 임베딩 지원

# Utils
python-dotenv
requests
pandas
numpy
